+++ hadoop-tools/hadoop-azure/src/test/java/org/apache/hadoop/fs/azurebfs/services/TestAbfsInputStream.java
@@ -85,11 +85,22 @@ public class TestAbfsInputStream extends
   private static final int INCREASED_READ_BUFFER_AGE_THRESHOLD =
       REDUCED_READ_BUFFER_AGE_THRESHOLD * 10; // 30 sec
   private static final int ALWAYS_READ_BUFFER_SIZE_TEST_FILE_SIZE = 16 * ONE_MB;
+  public static ReadBufferManager bufferManager = null;
+  
+  public TestAbfsInputStream() throws Exception {
+    super();
+    // Reduce thresholdAgeMilliseconds to 3 sec for the tests
+    if (getConfiguration().isReadAheadV2Enabled()) {
+      bufferManager = ReadBufferManagerV2.getBufferManager(getConfiguration());
+    } else {
+      bufferManager = ReadBufferManagerV1.getBufferManager();
+    }
+  }
 
   @Override
   public void teardown() throws Exception {
     super.teardown();
-    ReadBufferManager.getBufferManager().testResetReadBufferManager();
+    bufferManager.testResetReadBufferManager();
   }
 
   private AbfsRestOperation getMockRestOp() {
@@ -164,12 +175,12 @@ public class TestAbfsInputStream extends
 
   private void queueReadAheads(AbfsInputStream inputStream) {
     // Mimic AbfsInputStream readAhead queue requests
-    ReadBufferManager.getBufferManager()
+    bufferManager
         .queueReadAhead(inputStream, 0, ONE_KB, inputStream.getTracingContext());
-    ReadBufferManager.getBufferManager()
+    bufferManager
         .queueReadAhead(inputStream, ONE_KB, ONE_KB,
             inputStream.getTracingContext());
-    ReadBufferManager.getBufferManager()
+    bufferManager
         .queueReadAhead(inputStream, TWO_KB, TWO_KB,
             inputStream.getTracingContext());
   }
@@ -187,15 +198,15 @@ public class TestAbfsInputStream extends
   private void checkEvictedStatus(AbfsInputStream inputStream, int position, boolean expectedToThrowException)
       throws Exception {
     // Sleep for the eviction threshold time
-    Thread.sleep(ReadBufferManager.getBufferManager().getThresholdAgeMilliseconds() + 1000);
+    Thread.sleep(bufferManager.getThresholdAgeMilliseconds() + 1000);
 
     // Eviction is done only when AbfsInputStream tries to queue new items.
     // 1 tryEvict will remove 1 eligible item. To ensure that the current test buffer
     // will get evicted (considering there could be other tests running in parallel),
     // call tryEvict for the number of items that are there in completedReadList.
-    int numOfCompletedReadListItems = ReadBufferManager.getBufferManager().getCompletedReadListSize();
+    int numOfCompletedReadListItems = bufferManager.getCompletedReadListSize();
     while (numOfCompletedReadListItems > 0) {
-      ReadBufferManager.getBufferManager().callTryEvict();
+      bufferManager.callTryEvict();
       numOfCompletedReadListItems--;
     }
 
@@ -205,12 +216,6 @@ public class TestAbfsInputStream extends
     } else {
       inputStream.read(position, new byte[ONE_KB], 0, ONE_KB);
     }
-  }
-
-  public TestAbfsInputStream() throws Exception {
-    super();
-    // Reduce thresholdAgeMilliseconds to 3 sec for the tests
-    ReadBufferManager.getBufferManager().setThresholdAgeMilliseconds(REDUCED_READ_BUFFER_AGE_THRESHOLD);
   }
 
   private void writeBufferToNewFile(Path testFile, byte[] buffer) throws IOException {
@@ -364,7 +369,7 @@ public class TestAbfsInputStream extends
   public void testFailedReadAheadEviction() throws Exception {
     AbfsClient client = getMockAbfsClient();
     AbfsRestOperation successOp = getMockRestOp();
-    ReadBufferManager.setThresholdAgeMilliseconds(INCREASED_READ_BUFFER_AGE_THRESHOLD);
+    bufferManager.setThresholdAgeMilliseconds(INCREASED_READ_BUFFER_AGE_THRESHOLD);
     // Stub :
     // Read request leads to 3 readahead calls: Fail all 3 readahead-client.read()
     // Actual read request fails with the failure in readahead thread
@@ -379,7 +384,7 @@ public class TestAbfsInputStream extends
     // Add a failed buffer to completed queue and set to no free buffers to read ahead.
     ReadBuffer buff = new ReadBuffer();
     buff.setStatus(ReadBufferStatus.READ_FAILED);
-    ReadBufferManager.getBufferManager().testMimicFullUseAndAddFailedBuffer(buff);
+    bufferManager.testMimicFullUseAndAddFailedBuffer(buff);
 
     // if read failed buffer eviction is tagged as a valid eviction, it will lead to
     // wrong assumption of queue logic that a buffer is freed up and can lead to :
@@ -387,7 +392,7 @@ public class TestAbfsInputStream extends
     // at java.util.Stack.peek(Stack.java:102)
     // at java.util.Stack.pop(Stack.java:84)
     // at org.apache.hadoop.fs.azurebfs.services.ReadBufferManager.queueReadAhead
-    ReadBufferManager.getBufferManager().queueReadAhead(inputStream, 0, ONE_KB,
+    bufferManager.queueReadAhead(inputStream, 0, ONE_KB,
         getTestTracingContext(getFileSystem(), true));
   }
 
@@ -429,7 +434,7 @@ public class TestAbfsInputStream extends
     verifyReadCallCount(client, 3);
 
     // Sleep for thresholdAgeMs so that the read ahead buffer qualifies for being old.
-    Thread.sleep(ReadBufferManager.getBufferManager().getThresholdAgeMilliseconds());
+    Thread.sleep(bufferManager.getThresholdAgeMilliseconds());
 
     // Second read request should retry the read (and not issue any new readaheads)
     inputStream.read(ONE_KB, new byte[ONE_KB], 0, ONE_KB);
@@ -474,7 +479,7 @@ public class TestAbfsInputStream extends
             any(String.class), any(), any(TracingContext.class));
 
     AbfsInputStream inputStream = getAbfsInputStream(client, "testSuccessfulReadAhead.txt");
-    int beforeReadCompletedListSize = ReadBufferManager.getBufferManager().getCompletedReadListSize();
+    int beforeReadCompletedListSize = bufferManager.getCompletedReadListSize();
 
     // First read request that triggers readAheads.
     inputStream.read(new byte[ONE_KB]);
@@ -482,7 +487,7 @@ public class TestAbfsInputStream extends
     // Only the 3 readAhead threads should have triggered client.read
     verifyReadCallCount(client, 3);
     int newAdditionsToCompletedRead =
-        ReadBufferManager.getBufferManager().getCompletedReadListSize()
+        bufferManager.getCompletedReadListSize()
             - beforeReadCompletedListSize;
     // read buffer might be dumped if the ReadBufferManager getblock preceded
     // the action of buffer being picked for reading from readaheadqueue, so that
@@ -529,8 +534,7 @@ public class TestAbfsInputStream extends
             any(String.class), nullable(ContextEncryptionAdapter.class),
             any(TracingContext.class));
 
-    final ReadBufferManager readBufferManager
-        = ReadBufferManager.getBufferManager();
+    final ReadBufferManager readBufferManager = bufferManager;
 
     final int readBufferTotal = readBufferManager.getNumBuffers();
     final int expectedFreeListBufferCount = readBufferTotal
@@ -607,7 +611,7 @@ public class TestAbfsInputStream extends
     // if readAhead failed for specific offset, getBlock should
     // throw exception from the ReadBuffer that failed within last thresholdAgeMilliseconds sec
     intercept(IOException.class,
-        () -> ReadBufferManager.getBufferManager().getBlock(
+        () -> bufferManager.getBlock(
             inputStream,
             0,
             ONE_KB,
@@ -655,14 +659,14 @@ public class TestAbfsInputStream extends
     // AbfsInputStream Read would have waited for the read-ahead for the requested offset
     // as we are testing from ReadAheadManager directly, sleep for thresholdAgeMilliseconds so that
     // read buffer qualifies for to be an old buffer
-    Thread.sleep(ReadBufferManager.getBufferManager().getThresholdAgeMilliseconds());
+    Thread.sleep(bufferManager.getThresholdAgeMilliseconds());
 
     // Only the 3 readAhead threads should have triggered client.read
     verifyReadCallCount(client, 3);
 
     // getBlock from a new read request should return 0 if there is a failure
     // 30 sec before in read ahead buffer for respective offset.
-    int bytesRead = ReadBufferManager.getBufferManager().getBlock(
+    int bytesRead = bufferManager.getBlock(
         inputStream,
         ONE_KB,
         ONE_KB,
@@ -715,7 +719,7 @@ public class TestAbfsInputStream extends
     verifyReadCallCount(client, 3);
 
     // getBlock for a new read should return the buffer read-ahead
-    int bytesRead = ReadBufferManager.getBufferManager().getBlock(
+    int bytesRead = bufferManager.getBlock(
         inputStream,
         ONE_KB,
         ONE_KB,
@@ -853,7 +857,7 @@ public class TestAbfsInputStream extends
         .describedAs("Unexpected AlwaysReadBufferSize settings")
         .isEqualTo(alwaysReadBufferSizeEnabled);
 
-    Assertions.assertThat(ReadBufferManager.getBufferManager().getReadAheadBlockSize())
+    Assertions.assertThat(bufferManager.getReadAheadBlockSize())
         .describedAs("Unexpected readAhead block size")
         .isEqualTo(readAheadBlockSize);
 
@@ -921,8 +925,7 @@ public class TestAbfsInputStream extends
   }
 
   private void resetReadBufferManager(int bufferSize, int threshold) {
-    ReadBufferManager.getBufferManager()
-        .testResetReadBufferManager(bufferSize, threshold);
+    bufferManager.testResetReadBufferManager(bufferSize, threshold);
     // Trigger GC as aggressive recreation of ReadBufferManager buffers
     // by successive tests can lead to OOM based on the dev VM/machine capacity.
     System.gc();
